refactor : MockTestDataSetup 위치 수정
